+++ packages/video_player/video_player_avfoundation/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
+* Refactors native code for improved testing.
 
 ## 2.7.1